hw08_envdir_tool/executor.go
package main

import (
	"errors"
	"os"
	"os/exec"
)

// RunCmd runs a command + arguments (cmd) with environment variables from env.
func RunCmd(cmd []string, env Environment) (returnCode int) {
	command := exec.Command(cmd[0], cmd[1:]...) //nolint:gosec
	command.Stdin = os.Stdin
	command.Stdout = os.Stdout
	command.Stderr = os.Stderr

	environment := os.Environ()
	for key, value := range env {
		if value.NeedRemove {
			environment = append(environment, key+"=")
		} else {
			environment = append(environment, key+"="+value.Value)
		}
	}
	command.Env = environment

	if err := command.Run(); err != nil {
		var exitError *exec.ExitError
		if errors.As(err, &exitError) {
			return exitError.ExitCode()
		}
		return 1
	}

	return 0
}
